+++ Gems/Multiplayer/Code/Source/Components/LocalPredictionPlayerInputComponent.cpp
@@ -277,7 +277,7 @@ namespace Multiplayer
             input.SetClientInputId(GetLastInputId());
 
             ScopedAlterTime scopedTime(input.GetHostFrameId(), input.GetHostTimeMs(), input.GetHostBlendFactor(), invokingConnection->GetConnectionId());
-            GetNetBindComponent()->ProcessInput(input, clientInputRateSec);
+            GetNetBindComponent()->ProcessInput(input, static_cast<float>(clientInputRateSec));
 
             AZLOG(NET_Prediction, "Migrated InputId=%d", aznumeric_cast<int32_t>(input.GetClientInputId()));
 
@@ -345,7 +345,7 @@ namespace Multiplayer
             // Reprocess the input for this frame
             NetworkInput& input = m_inputHistory[replayIndex];
             ScopedAlterTime scopedTime(input.GetHostFrameId(), input.GetHostTimeMs(), input.GetHostBlendFactor(), invokingConnection->GetConnectionId());
-            GetNetBindComponent()->ReprocessInput(input, clientInputRateSec);
+            GetNetBindComponent()->ReprocessInput(input, static_cast<float>(clientInputRateSec));
 
             AZLOG(NET_Prediction, "Replayed InputId=%d", aznumeric_cast<int32_t>(input.GetClientInputId()));
         }
@@ -438,10 +438,10 @@ namespace Multiplayer
             input.SetHostBlendFactor(multiplayer->GetCurrentBlendFactor());
 
             // Allow components to form the input for this frame
-            GetNetBindComponent()->CreateInput(input, clientInputRateSec);
+            GetNetBindComponent()->CreateInput(input, static_cast<float>(clientInputRateSec));
 
             // Process the input for this frame
-            GetNetBindComponent()->ProcessInput(input, clientInputRateSec);
+            GetNetBindComponent()->ProcessInput(input, static_cast<float>(clientInputRateSec));
 
             AZLOG(NET_Prediction, "Processed InputId=%d", aznumeric_cast<int32_t>(m_clientInputId));
 
@@ -506,7 +506,7 @@ namespace Multiplayer
             NetworkInput& input = m_lastInputReceived[0];
             {
                 ScopedAlterTime scopedTime(input.GetHostFrameId(), input.GetHostTimeMs(), DefaultBlendFactor, GetNetBindComponent()->GetOwningConnectionId());
-                GetNetBindComponent()->ProcessInput(input, clientInputRateSec);
+                GetNetBindComponent()->ProcessInput(input, static_cast<float>(clientInputRateSec));
             }
 
             AZLOG(NET_Prediction, "Forced InputId=%d", aznumeric_cast<int32_t>(input.GetClientInputId()));
